fix: change `int` to `float` in ocr

+++ src/services/ocr.py
@@ -38,10 +38,10 @@ async def parse_nutrients_from_image(image: bytes) -> FoodContentOptional:
     # 총 내용량 파싱
     total_content = re.search(r"총내용량 ?([\d.]+) ?g", text)
     # total_content = re.search(r'총내용량([\d\.]+g(?:\([\d\.]+gX\d+봉지\))?)', text)
-    total_content = int(total_content.group(1)) if total_content else None
+    total_content = float(total_content.group(1)) if total_content else None
     if total_content is None:
         total_content = re.search(r"내용량 ?([\d.]+) ?g", text)
-        total_content = int(total_content.group(1)) if total_content else None
+        total_content = float(total_content.group(1)) if total_content else None
 
     # 단위 파싱
     per_unit = re.search(r"총 ?내용량 ?당", text)
@@ -66,15 +66,15 @@ async def parse_nutrients_from_image(image: bytes) -> FoodContentOptional:
 
     # 탄수화물 파싱
     carb = re.search(r"탄수화물([\d.,]+)g", text)
-    carb = int(carb.group(1)) if carb else None
+    carb = float(carb.group(1)) if carb else None
 
     # 당류 파싱
     sugar = re.search(r"당류([\d.,]+)g", text)
-    sugar = int(sugar.group(1)) if sugar else None
+    sugar = float(sugar.group(1)) if sugar else None
 
     # 지방 파싱
     fat = re.search(r"지방([\d.,]+)g", text)
-    fat = int(fat.group(1)) if fat else None
+    fat = float(fat.group(1)) if fat else None
 
     # # 트랜스지방 파싱
     # trans_fat = re.search(r"트랜스지방([\d\.,]+)g", text)
@@ -90,7 +90,7 @@ async def parse_nutrients_from_image(image: bytes) -> FoodContentOptional:
 
     # 단백질 파싱
     protein = re.search(r"단백질([\d.,]+)g", text)
-    protein = int(protein.group(1)) if protein else None
+    protein = float(protein.group(1)) if protein else None
 
     # # 칼슘 파싱
     # calcium = re.search(r"칼슘([\d\.,]+)mg", text)
